scripts/requirements.txt
requests
pandas
azure-storage-blob
schedule
python-dotenv